fix: remux.c の不要なキャストを削除する

+++ crates/c-api/examples/remux.c
@@ -287,7 +287,7 @@ int main(int argc, char *argv[]) {
             .data_offset = current_output_data_offset,
             .data_size = (uint32_t)demux_sample.data_size,
             .has_composition_time_offset = demux_sample.has_composition_time_offset,
-            .composition_time_offset = (int32_t)demux_sample.composition_time_offset,
+            .composition_time_offset = demux_sample.composition_time_offset,
         };
 
         // マルチプレックサーにサンプルを追加
